feat: QE 빠른 범위 30m 추가, 시작/종료 라벨 from/to 변경
- preset 버튼에 30m(data-h="0.5") 추가
- parseInt → parseFloat으로 변경 (0.5 정상 처리)
- 시작/종료 라벨을 from/to로 변경

+++ static/js/explorer.js
@@ -61,7 +61,7 @@ async function qeInit() {
   $('qe-add-cond').onclick   = qeAddCondRow;
 
   document.querySelectorAll('.preset-btn').forEach(btn => {
-    btn.onclick = () => qeSetPreset(parseInt(btn.dataset.h));
+    btn.onclick = () => qeSetPreset(parseFloat(btn.dataset.h));
   });
 
   document.querySelector('.filters').addEventListener('keydown', e => {
